+++ lib/main.dart
@@ -9,7 +9,7 @@ import 'package:hydrated_bloc/hydrated_bloc.dart';
 import 'cubits/index.dart';
 import 'repositories/index.dart';
 import 'services/index.dart';
-import 'util/routes.dart';
+import 'util/index.dart';
 
 Future<void> main() async {
   WidgetsFlutterBinding.ensureInitialized();
@@ -73,6 +73,16 @@ class CherryApp extends StatelessWidget {
     this.changelogRepository,
   });
 
+  /// Calls the `NotificationsCubit` to update the scheduled notifications
+  /// necessary
+  static void scheduleLaunchNotification(BuildContext context) =>
+      context.watch<NotificationsCubit>().updateNotifications(
+            context,
+            nextLaunch: LaunchUtils.getUpcomingLaunch(
+              context.watch<LaunchesCubit>().state.value,
+            ),
+          );
+
   @override
   Widget build(BuildContext context) {
     return MultiBlocProvider(
@@ -88,42 +98,25 @@ class CherryApp extends StatelessWidget {
       ],
       child: BlocConsumer<ThemeCubit, ThemeState>(
         listener: (context, state) => null,
-        builder: (context, state) => MaterialApp(
-          title: 'SpaceX GO!',
-          theme: context.watch<ThemeCubit>().lightTheme,
-          darkTheme: context.watch<ThemeCubit>().darkTheme,
-          themeMode: context.watch<ThemeCubit>().themeMode,
-          onGenerateRoute: Routes.generateRoute,
-          onUnknownRoute: Routes.errorRoute,
-          localizationsDelegates: [
-            FlutterI18nDelegate(
-              translationLoader: FileTranslationLoader(),
-            )..load(null),
-            GlobalMaterialLocalizations.delegate,
-            GlobalWidgetsLocalizations.delegate
-          ],
-        ),
+        builder: (context, state) {
+          scheduleLaunchNotification(context);
+          return MaterialApp(
+            title: 'SpaceX GO!',
+            theme: context.watch<ThemeCubit>().lightTheme,
+            darkTheme: context.watch<ThemeCubit>().darkTheme,
+            themeMode: context.watch<ThemeCubit>().themeMode,
+            onGenerateRoute: Routes.generateRoute,
+            onUnknownRoute: Routes.errorRoute,
+            localizationsDelegates: [
+              FlutterI18nDelegate(
+                translationLoader: FileTranslationLoader(),
+              )..load(null),
+              GlobalMaterialLocalizations.delegate,
+              GlobalWidgetsLocalizations.delegate
+            ],
+          );
+        },
       ),
     );
-  }
-}
-
-class CherryBlocObserver extends BlocObserver {
-  @override
-  void onCreate(Cubit cubit) {
-    super.onCreate(cubit);
-    debugPrint('onCreate: ${cubit.runtimeType}');
-  }
-
-  @override
-  void onChange(Cubit cubit, Change change) {
-    super.onChange(cubit, change);
-    debugPrint('onChange: ${cubit.runtimeType}, $change');
-  }
-
-  @override
-  void onError(Cubit cubit, Object error, StackTrace stackTrace) {
-    super.onError(cubit, error, stackTrace);
-    debugPrint('onError: ${cubit.runtimeType}, $error');
   }
 }
