+++ lib/ui/screens/start.dart
@@ -1,11 +1,9 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:flutter_i18n/flutter_i18n.dart';
 import 'package:flutter_svg/flutter_svg.dart';
 import 'package:google_fonts/google_fonts.dart';
 import 'package:quick_actions/quick_actions.dart';
 
-import '../../cubits/index.dart';
 import '../tabs/index.dart';
 
 /// This view holds all tabs & its models: home, vehicles, upcoming & latest launches, & company tabs.
@@ -74,13 +72,6 @@ class _StartScreenState extends State<StartScreen> {
 
   @override
   Widget build(BuildContext context) {
-    // Calls the `NotificationsCubit` to update the scheduled notifications
-    // necessary
-    // context.watch<NotificationsCubit>().updateNotifications(
-    //       context,
-    //       nextLaunch: context.watch<LaunchesCubit>().state.value,
-    //     );
-
     return Scaffold(
       body: IndexedStack(index: _currentIndex, children: [
         HomeTab(),
