+++ lib/util/bloc_observer.dart
@@ -1,0 +1,22 @@
+import 'package:flutter/foundation.dart';
+import 'package:flutter_bloc/flutter_bloc.dart';
+
+class CherryBlocObserver extends BlocObserver {
+  @override
+  void onCreate(Cubit cubit) {
+    super.onCreate(cubit);
+    debugPrint('onCreate: ${cubit.runtimeType}');
+  }
+
+  @override
+  void onChange(Cubit cubit, Change change) {
+    super.onChange(cubit, change);
+    debugPrint('onChange: ${cubit.runtimeType}, $change');
+  }
+
+  @override
+  void onError(Cubit cubit, Object error, StackTrace stackTrace) {
+    super.onError(cubit, error, stackTrace);
+    debugPrint('onError: ${cubit.runtimeType}, $error');
+  }
+}
