+++ lib/util/index.dart
@@ -1,4 +1,5 @@
 export 'api_query.dart';
+export 'bloc_observer.dart';
 export 'browser.dart';
 export 'colors.dart';
 export 'launch_utils.dart';
