+++ lib/util/launch_utils.dart
@@ -5,12 +5,20 @@ class LaunchUtils {
   /// Returs the most upcoming launch inside a list, if the list
   /// is sorted by date.
   static Launch getUpcomingLaunch(List<Launch> launches) {
-    return launches.reversed.where((l) => l.upcoming).first;
+    if (launches != null) {
+      return launches.reversed.where((l) => l.upcoming).first;
+    } else {
+      return null;
+    }
   }
 
   /// Returs the most latest launch inside a list, if the list
   /// is sorted by date.
   static Launch getLatestLaunch(List<Launch> launches) {
-    return launches.where((l) => !l.upcoming).first;
+    if (launches != null) {
+      return launches.where((l) => !l.upcoming).first;
+    } else {
+      return null;
+    }
   }
 }
